[FIX] hr: last login hour on employee kanban.
Setting a value for the `date` field on a partner of
an employee leaded to a crash in the employees
kanban view.

This last login field on the employee kanban
had as goal to display the last connection hour
of each employee, if the last
login happened within the current day.

Since 7.0, the last login date field  on `res.users`
has been renamed from `date` to `login_date`,
and became a date instead of a datetime.

Nevertheless, the `last_login` related field
on `hr.employee` remained on the field `date`,
which is set on the `res.partner` model,
but is not used as a connection date field. The
related should have been updated to use
`res.users`.`login_date`, probably, but it
wasn't the case.

From saas-6, the `date` field is no longer available
on the `res.partner` form.

We can therefore safely assume that this feature
is no longer used nor needed, as:
 - It should display the last hour of connection within a day
 - but it uses the wrong date field
 - and it uses a field type for which the hour is not stored (date).
 - even by updating the related to use `login_date`,
   it would still be a field where the hour is not stored.

opw-654313