[FIX] pos_mercury: make forcing validation work again
The POS will asks the user to confirm certain things that could be a
user error. Examples of this are attempting to validate an order with a
very big payment compared to the amount to pay on the order. Another
example is paying an amount of 0 for an order with a total amount to pay
that is negative (manual refund).

The issue was that pos_mercury overwrote validate_order(), and didn't
call _super() correctly, losing the user confirmation for these type of
actions.