[IMP] ir_translation: make access rights depend on translated record
A translation of a field value for a given record can only be created/updated/
deleted if the current user has access rights for updating the record.

Access rights on translations are restricted to the group "Employee".  Those
still need create/write/delete access, in order to be able to edit records in
the user interface.  Other users (portal, etc.) simply have no explicit access
rights at all on translations.