[FIX] translate: fix `xml_translate` to use an HTML parser as a fallback
When terms are translated in the website, their values are sent as HTML (with
text entities escaped).  The function `xml_translate` is used to sanitize those
values, and sometimes crashes because of non-closed entities like `<br>`.  Fix
it by using an HTML parser as a fallback.

This is a partial revert of f90b4a75f2927b3fba84e4956c09452ff841692b.