[FIX] project: IE11+ issue with flex and overflow
IE11 and IE Edge presented issue in the project kanban view with content
overflowing its container.

On IE Edge it could be solved by "min-width:0" on the left part of a
kanban card, so it was an issue in the computation of widths order.

But on IE11 that wasn't enough so a less elegant fix has been used.

opw-653477

(solution found mainly by qsm)