[FIX] stock_dropshipping: do not raise a warning for dropshipping product
When selling a product with a dropshipping route, the available stock was
checked and a warning got raise if it was the case.
The _check_routing was not working properly as is called in a context of an
onchange and 'self' was the record in the database, not in the browser.

Just use check there is a dropshipping route on the product instead of checking
it on the routes of the sale order line (not reliable info anyway).

Add a todo for master to improve it to change the signature of the method.

opw 657760