[FIX] website_mass_mailing: edit snippet newsletter
The "edit" action is on the "website.TopBar", which will create a new
instance of the "editor" (web_editor.editor) widget.

So if we want to do something when the "editing" start, we have to
either override "edit" of the TopBar, or override "start" or "init" of
the editor.

opw-657662