[FIX] mail: performance in find_partner_from_email
This is a backport of commit 79a375b2c744cb8d8b84304ca2facdd0eefa5bd0 done
in 9 and ported to the old API. The purpose is to avoid searching on
the res.partner table with a domain leaf being user_ids != False. Indeed
this search is costly. Use a direct search on the user table instead.