[FIX] point_of_sale: make tendered amount adhere to currency formatting
Part of the number formatting overhaul done in
cdba7987c0bdb27379526c0af196dbd6261a484f,
9a653060335af56969c106d409d08e4d29eb445a,
bc777804f1b9a23807446f12398015648eb61d4d and
9b1e6c2ccc665f88613d8ecc02fa94b15fce09bb. The only place left where we
didn't correctly format the numbers was in the tendered column on
the payment screen. This fixes that by not directly displaying the
inputbuffer anymore. Instead we pass the payment line amount through
format_currency_no_symbol() and display that instead. We also give a
default amount of 0 to parse_value so it doesn't return false for an
empty string.