[FIX] res.lang: do not leak `load_language` global
During database creation via the database manager,
or when using the startup option `--load-language`,
the selected language(s) will be installed as soon
as either:
 - the base module is installed/updated (because
base_data.xml includes a call to res.lang.install_lang()
 - the registry is loaded (after loading `base`,
the system installs the requested languages, even
if the server is not in update/install mode)

This is implemented by passing a global config
option: `load_lang`.

This behavior was modified as of saas-7 by PR
for the command-line and for the database manager.

In both cases, we don't want the installation
to be repeated the next time either of these
event occur. Essentially the `load_language`
config option is a single-use flag, so we
have to pop it from the config after using it.

A more complete refactoring is needed to
remove all uses or tools.config as a
global "context" for passing parameters.